kernel: Handle additional flags earlier
Currently, it is possible that some flags aren't used in specific
stages of kernel compilation (e.g. headers_install) because these
flags aren't added to KERNEL_MAKE_FLAGS early enough for exporting
as a soong var, causing tasks utilizing the var to not compile
with the expected flags.

This change mitigates the problem by moving the appending of the
flags to a position before exporting the flags to soong.

Change-Id: I969fc30d1967c661fe7c4ef3c8a261088947851d
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>